refactoring email
